refactor: refactored column config
